screen-context/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0

# Screen capture
pillow>=10.2.0
mss>=9.0.1

# OpenAI for vision analysis
openai>=1.12.0

# macOS specific (optional, for better accessibility)
pyobjc-core>=10.1;sys_platform=="darwin"
pyobjc-framework-Quartz>=10.1;sys_platform=="darwin"
pyobjc-framework-ApplicationServices>=10.1;sys_platform=="darwin"
